split out build part of manager
