+++ launch_chffrplus.sh
@@ -220,7 +220,7 @@ function launch {
 
   # start manager
   cd selfdrive/manager
-  ./manager.py
+  ./build.py && ./manager.py
 
   # if broken, keep on screen error
   while true; do sleep 1; done
