+++ release/files_common
@@ -385,10 +385,11 @@ selfdrive/camerad/imgproc/utils.cc
 selfdrive/camerad/imgproc/utils.h
 
 selfdrive/manager/__init__.py
+selfdrive/manager/build.py
 selfdrive/manager/helpers.py
-selfdrive/manager/process.py
-selfdrive/manager/process_config.py
 selfdrive/manager/manager.py
+selfdrive/manager/process_config.py
+selfdrive/manager/process.py
 
 selfdrive/modeld/SConscript
 selfdrive/modeld/modeld.cc
